Update match model

+++ frontend/src/models/match_model.rs
@@ -2,7 +2,7 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Default, Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Match {
-    pub(crate) id: u32,
+    pub(crate) id: u64,
     pub(crate) date: u64,
 
     #[serde(rename = "knockoutStage")]
